fix: AttributeErrors following last commit
# Issue
Following last commit of sprites.py, which changed most attributes in FallObject sprite to 'private' attribute, raised AttributeError when running the simulation due to some functions use/modify attributes directly from object, which is not possible in last commit.
# Solution
One possible solution is to remove the prepending '__'. But I want to try something new. So I use @property decorator instead for getting and setting the values of position and speed attribute.